Create a profile with Image
Instead of having a separate page for creating your profile, I'm suggesting that we should instead direct new users to edit-profile-page. This means that, visually, we aren't providing users with different experiences.

For this to happen, I've added to the check we make during login (to see if we have an entry for them in the database). If the user doesn’t exist, we direct them to the edit-profile-page. In this page, I’ve added image picking with square cropping and quality reduction so that upload size isn’t too big. This also means that image downloads will be faster and consume less data.